requirements.txt
click==8.1.7
pyyaml==6.0.1
markdown==3.4.3
pytest>=7.0.0
black==23.11.0
flake8==6.1.0
pytest-cov>=4.0.0
python-semantic-release==8.7.0
coverage>=7.0.0
